Flow Plot + Major Plotting Refactor (#283)
This change adds a flow plot with values calculated from bucket tip events that are binned every 90 minutes. Major refactors of the plotting infrastructure are also included to create more general Plot and PlotCard components that are reused across both temperature and flow views. This closes #134 and closes #242. There is still some cleanup and refactoring that will be left for a future diff (like #282 and #198).