SECURITY FIX: CVE-2021-24033
Vulnerable versions: >= 0.4.0, < 11.0.4
Patched version: 11.0.4

react-dev-utils prior to v11.0.4 exposes a function, getProcessForPort,
where an input argument is concatenated into a command string to be
executed. This function is typically used from react-scripts (in Create
React App projects), where the usage is safe. Only when this function
is manually invoked with user-provided values (ie: by custom code) is
there the potential for command injection. If you're consuming it from
react-scripts then this issue does not affect you.